main_svc 5

+++ stats/dto/src/main/java/ru/practicum/dto/ViewStats.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class ViewStats {
     private String app;
     private String uri;
-    private Long hits;
+    private Integer hits;
 }
